+++ test/date.spec.ts
@@ -1,6 +1,6 @@
 import assert from 'assert';
 import { format } from 'date-fns';
-import { dayDiff, today, isValidDate } from '../dist';
+import { dayDiff, today, isValidDate, dateToYYYYMMDD } from '../dist';
 
 describe('Date', () => {
 	it('dayDiff', (done) => {
@@ -22,6 +22,12 @@ describe('Date', () => {
 		assert.strictEqual(isValidDate('2021-02-29'), false);
 		assert.strictEqual(isValidDate('2021-03-32'), false);
 		assert.strictEqual(isValidDate('2021-13-01'), false);
+		done();
+	});
+
+	it('dateToYYYYMMDD', (done) => {
+		assert.strictEqual(dateToYYYYMMDD(new Date('2023-05-15T01:01:00Z')), '2023-05-15');
+		assert.strictEqual(dateToYYYYMMDD(new Date(2023, 11, 31), '/'), '2023/12/31');
 		done();
 	});
 });