+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.6 (2023-02-)
 
 - `isValidDate`: Only the `yyyy-mm-dd` format can be verified
+- `dateToYYYYMMDD`: Add `dateToYYYYMMDD` method
 
 ## 1.1.5 (2023-02-07)